Stabilize GRPO reward normalization

+++ miles/ray/rollout/train_data_conversion.py
@@ -113,7 +113,10 @@ def _post_process_rewards(args, samples: list[Sample] | list[list[Sample]], cust
     raw_rewards = [sample.get_reward_value(args) for sample in samples]
     if args.advantage_estimator in ["grpo", "gspo", "reinforce_plus_plus_baseline"] and args.rewards_normalization:
         # group norm
-        rewards = torch.tensor(raw_rewards, dtype=torch.float)
+        # Center in float64. Float32 reduction can move the mean of an identical
+        # non-representable reward (for example, eight 0.9 values) by one ULP,
+        # which the GRPO epsilon then amplifies into a spurious policy signal.
+        rewards = torch.tensor(raw_rewards, dtype=torch.float64)
         if rewards.shape[-1] == args.n_samples_per_prompt * args.rollout_batch_size:
             rewards = rewards.reshape(-1, args.n_samples_per_prompt)
         else:
